+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#7](https://github.com/LaxarJS/ax-details-layer-widget/issues/7): fixed CSS class names and copyright holder
+    + **BREAKING CHANGE:** see ticket for details
 - [#8](https://github.com/LaxarJS/ax-details-layer-widget/issues/8): iOS: prevent background scrolling on mobile safari